fix(unidade-producao/data-access): DbType.Int32 to DbType.Int64
Muda a conversão da coluna PROPRIEDADE_CODIGO de Int32 para Int64 em UnidadeProducaoDa.cs

+++ Código Fonte/VERSAO-ATUAL/src/EtramiteX/Model.Extencao/Caracterizacoes/ModuloUnidadeProducao/Data/UnidadeProducaoDa.cs
@@ -83,7 +83,7 @@ namespace Tecnomapas.EtramiteX.Interno.Model.Extensoes.Caracterizacoes.ModuloUni
 
 				comando.AdicionarParametroEntrada("empreendimento_id", caracterizacao.Empreendimento.Id, DbType.Int32);
 				comando.AdicionarParametroEntrada("possui_cod_propriedade", caracterizacao.PossuiCodigoPropriedade, DbType.Int32);
-				comando.AdicionarParametroEntrada("propriedade_codigo", caracterizacao.CodigoPropriedade, DbType.Int32);
+				comando.AdicionarParametroEntrada("propriedade_codigo", caracterizacao.CodigoPropriedade, DbType.Int64);
 				comando.AdicionarParametroEntrada("local_livro", caracterizacao.LocalLivroDisponivel, DbType.String);
 				comando.AdicionarParametroEntrada("tid", DbType.String, 36, GerenciadorTransacao.ObterIDAtual());
 				comando.AdicionarParametroSaida("id", DbType.Int32);
